requirements.txt
aiogram>=3.4.0
openai>=1.12.0
anthropic>=0.8.0
openai-whisper
whisperx
pyannote.audio
torch
torchaudio
pydantic>=2.4.0,<2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.26.0
jinja2>=3.1.0
aiosqlite>=0.19.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
psutil>=5.9.0
aiofiles>=23.0.0
